move helpers into a separate library
fixes #34
